requirements.txt
pyserial==3.5
git+https://github.com/ppapryczka/FaultInjectionMessages.git#egg=fault-injection-messages
